requirements.txt
aksharamukha==2.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
assemblyai==0.44.3
audioread==3.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
decorator==5.2.1
deep-translator==1.11.4
Deprecated==1.2.18
exceptiongroup==1.3.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==6.0.1
marisa-trie==1.3.1
MarkupSafe==3.0.2
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai-whisper==20250625
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
proglog==0.1.12
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
pykakasi==2.3.0
python-dotenv==1.1.1
pytube==15.0.0
PyYAML==6.0.2
regex==2025.9.1
reportlab==4.0.8
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
unicodedata2==16.0.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
